Create EvaluatorModelView to factorize code
